ai_agent_service/requirements.txt
flask>=2.3.0
groq==0.9.0
python-dotenv>=1.0.0
gunicorn==23.0.0
